add fields to "create session" and "add user"

+++ src/controllers/session.controller.js
@@ -1,48 +1,44 @@
 const httpStatus = require('http-status');
 const catchAsync = require('../utils/catchAsync');
-const { sessionService } = require('../services');
+const { sessionService, emailService } = require('../services');
 
 const createSession = catchAsync(async (req, res) => {
-  const { sessionName, startTime, startDate, duration, email} = req.body;
+  const { sessionName, notaryField, notaryService, startTime, startDate, duration, email } = req.body;
   const createdBy = req.user.id;
-  const checkemail = await sessionService.validateEmails(email);
+  await sessionService.validateEmails(email);
   const session = await sessionService.createSession({
     sessionName,
+    notaryField,
+    notaryService,
     startTime: new Date(startTime),
     startDate: new Date(startDate),
     duration,
-    email,  
-    createdBy,  
+    email,
+    createdBy,
   });
+  await Promise.all(session.email.map((emailItem) => emailService.sendInvitationEmail(emailItem, session._id)));
   res.status(httpStatus.CREATED).send(session);
 });
 
-
 const addUserToSession = catchAsync(async (req, res) => {
-    const sessionId = req.params.sessionId;
-    const { email } = req.body;
-    const session = await sessionService.findBySessionId(sessionId);
-    const checkemail = await sessionService.validateEmails(email);
-    const updatedSession = await sessionService.addUserToSession({
-      sessionId,
-      email,
-    });
-    res.status(httpStatus.OK).send(updatedSession);
+  const { sessionId } = req.params;
+  const { email } = req.body;
+  const userId = req.user.id;
+  const updatedSession = await sessionService.addUserToSession({ sessionId, email, userId });
+  await Promise.all(
+    updatedSession.email.map((emailItem) => emailService.sendInvitationEmail(emailItem, updatedSession._id))
+  );
+  res.status(httpStatus.OK).send(updatedSession);
 });
-
 
 const deleteUserOutOfSession = catchAsync(async (req, res) => {
-    const sessionId = req.params.sessionId;
-    const { email } = req.body;
-    const session = await sessionService.findBySessionId(sessionId);
-    const checkemail = await sessionService.validateEmails(email);
-    const updatedSession = await sessionService.deleteUserOutOfSession({
-      sessionId,
-      email,
-    });
-    res.status(httpStatus.OK).send(updatedSession);
+  const { sessionId } = req.params;
+  const { email } = req.body;
+  const userId = req.user.id;
+  const updatedSession = await sessionService.deleteUserOutOfSession({ sessionId, email, userId });
+  res.status(httpStatus.OK).send(updatedSession);
 });
-  
+
 module.exports = {
   createSession,
   addUserToSession,
